moved token related CONSTANTS to token folder
