+++ src/tokens/token-lists/KOVAN/prod-token-list.json
@@ -1,0 +1,44 @@
+{
+  "elements": [
+    {
+      "symbol": "ETH",
+      "name": "Ether",
+      "address": "0x0",
+      "decimals": 18,
+      "isETH": true
+    },
+    {
+      "symbol": "RDN",
+      "name": "Raiden Token",
+      "address": "0x1f7f270df126ba464228cc8d8203d2768429e085",
+      "decimals": 18
+    },
+    {
+      "symbol": "WETH",
+      "name": "Wrapped Ether",
+      "address": "0xd0a1e359811322d97991e03f863a0c30c2cf029c",
+      "decimals": 18
+    },
+    {
+      "symbol": "OMG",
+      "name": "OmiseGO",
+      "address": "0xde6efd396e18a950b45e24d6225505f48d0c627b",
+      "decimals": 18
+    }
+  ],
+  "pagination": {
+    "endingBefore": null,
+    "startingAfter": null,
+    "limit": 20,
+    "order": [
+      {
+        "param": "symbol",
+        "direction": "ASC"
+      }
+    ],
+    "previousUri": null,
+    "nextUri": null
+  },
+  "network": "KOVAN",
+  "version": "1.0.1"
+}
