+++ src/utils/index.ts
@@ -7,7 +7,8 @@ import { store } from 'components/App'
 import { getNetwork } from 'integrations/provider'
 import { DefaultTokenList, ProviderInterface, DefaultTokenObject, Receipt, ABI, Web3EventLog, TransactionObject } from 'api/types'
 import { Account } from 'types'
-import { ETH_ADDRESS, WALLET_PROVIDER, DEFAULT_ERROR, CANCEL_TX_ERROR, NO_INTERNET_TX_ERROR, LOW_GAS_ERROR, ProviderName, ProviderType, GAS_LIMIT, URLS, BLOCKED_COUNTRIES } from 'globals'
+import { WALLET_PROVIDER, DEFAULT_ERROR, CANCEL_TX_ERROR, NO_INTERNET_TX_ERROR, LOW_GAS_ERROR, ProviderName, ProviderType, GAS_LIMIT, URLS, BLOCKED_COUNTRIES } from 'globals'
+import { ETH_ADDRESS } from 'tokens'
 
 import GNOSIS_SAFE_SVG from 'assets/img/icons/icon_gnosis_safe1.svg'
 import STATUS_SVG from 'assets/img/icons/icon_status.svg'
